add email template for better email view and modified passport for send more data to user

+++ Passport-services/api/service/passport/passport.js
@@ -5,7 +5,8 @@ const e = require("express");
 
 exports.createPassport = async (data, email) => {
   const passport = await dal.create(passportModel, data);
-  sendMessage("passport.created", "Passport created", email);
+
+  sendMessage("passport.created", "Passport created", email,passport);
 
   return {
     status: 200,
@@ -24,12 +25,14 @@ exports.updatePassport = async (id, data, email) => {
   const fingpassport = await dal.findByID(passportModel, id);
   if (!fingpassport) return { status: 404, message: "Passport not found" };
   const passport = await dal.findOneAndUpdate(passportModel, { _id: id }, data);
-  sendMessage("passport.updated", "Passport updated", email);
+  sendMessage("passport.updated", "Passport updated", email,passport);
   return { status: 200, message: "Passport updated", data: passport };
 };
 
 exports.deletePassport = async (id, email) => {
-  const passport = await dal.findOneAndDelete(passportModel, { _id: id });
-  sendMessage("passport.deleted", "Passport deleted", email);
-  return { status: 200, message: "Passport deleted" };
+  const passportData = await dal.findByID(passportModel, id);
+  if (!passportData) return { status: 404, message: "Passport not found" };
+   await dal.findOneAndDelete(passportModel, { _id: id });
+  sendMessage("passport.deleted", "Passport deleted", email, { data: `Passport is deleted with ID: ${passportData._id.toString()}` });
+  return { status: 200, message: "Passport deleted", data: id };
 };
